Fix Auto Recorder plugin to replace slashes in the map name with dashes.
When using the plugin in CSGO, Auto Recorder will fail to open a demo file for games played on workshop maps. In CSGO, workshop maps are stored in a directory hierarchy, and are identified with a name following following the pattern "workshop/<map_id>/<map_name>.bsp". Auto Recorder would attempt to create demos by concatenating the date with the map name, resulting in the plugin trying to create a demo file in a folder that does not exist. Replacing the slashes with dashes fixed the problem and creates a flat demo folder structure.